requirements.txt
Django==3.2
django-cors-headers==3.12.0
djangorestframework==3.13.1
